Nodes CRUD Finished

+++ src/main/java/com/stakroute/neo4j/service/Service.java
@@ -5,16 +5,17 @@ import com.stakroute.neo4j.model.Node;
 
 import java.util.List;
 
+
 public interface Service {
     List<Node> getAll();
 
     Node save(Node node);
 
-    Node delete(int id);
+    Node delete(long id);
 
     //Node deleteByName(String name);
     //Node deleteById(int id);
     //Node deleteByParent(String parent);
 
-    Node update(int id, String name, String parent);
+    Node update(long id, String name, String parent);
 }
